Change all u32 to u64

+++ src/main.rs
@@ -1,7 +1,7 @@
 use std::{env::args, process::exit, num::IntErrorKind, time::{Duration, Instant}};
 
 // Print factors and time elapsed
-fn print_results(factors: &Vec<u32>, duration: Duration) {
+fn print_results(factors: &Vec<u64>, duration: Duration) {
     if duration.as_secs() > 0 && duration.subsec_millis() > 0 { print!("[{}.{}s] ", duration.as_secs(), duration.subsec_millis()/10); }
     if factors.len() > 0 { for factor in factors { print!("{} ", factor); } }
     else { print!("1"); }
@@ -9,8 +9,8 @@ fn print_results(factors: &Vec<u32>, duration: Duration) {
 }
 
 // Find factors and duration
-fn factor_arg(mut arg: u32) -> (Vec<u32>, Duration) {
-    let mut factors: Vec<u32> = vec![];
+fn factor_arg(mut arg: u64) -> (Vec<u64>, Duration) {
+    let mut factors: Vec<u64> = vec![];
 
     let now = Instant::now();
 
@@ -25,7 +25,7 @@ fn factor_arg(mut arg: u32) -> (Vec<u32>, Duration) {
 }
 
 // Parse input to number
-fn parse_arg(mut arg: String) -> u32 {
+fn parse_arg(mut arg: String) -> u64 {
     arg = if arg.contains('-') { arg.replace('-', "") } else { arg };
     arg = if arg.contains(',') { arg.replace(',', "") } else { arg };
     arg = if arg.contains('.') { arg.replace('.', "") } else { arg };
@@ -36,7 +36,7 @@ fn parse_arg(mut arg: String) -> u32 {
         Err(err) => {
             match err.kind() {
                 IntErrorKind::InvalidDigit => { eprintln!("Error: Only use (0-9), comma, underscore, or period"); exit(2); },
-                IntErrorKind::PosOverflow => { eprintln!("Error: Integer too large\nMaxInt: {}", u32::MAX); exit(2); }
+                IntErrorKind::PosOverflow => { eprintln!("Error: Integer too large\nMaxInt: {}", u64::MAX); exit(2); }
                 IntErrorKind::Empty => { eprintln!("Error: Must include numbers (0-9)"); exit(2); }
                 _ => { eprintln!("Error: {}", err); exit(2); }
             }
@@ -58,7 +58,7 @@ fn main() {
     // println!("String: {}", arg_str);
 
     let arg = parse_arg(arg);
-    // println!("u32: {}", arg);
+    // println!("u64: {}", arg);
 
     let (factors, duration) = factor_arg(arg);
     // println!("{:?}", factors);
